users

+++ src/public/scripts/users.js
@@ -6,29 +6,40 @@ displayUsers();
 hidenav();
 
 function displayUsers() {
-    Http.Get("/api/users/all")
-        .then((response) => response.json())
-        .then((response) => {
-            var allUsers = response.users;
-            // Empty the anchor
-            var allUsersAnchor = document.getElementById("all-users-anchor");
-            allUsersAnchor.innerHTML = "";
-            // Append users to anchor
-            allUsers.forEach((user) => {
-                allUsersAnchor.innerHTML += getUserDisplayEle(user);
-            });
-        });
+  Http.Get("/api/users/all")
+    .then((response) => response.json())
+    .then((response) => {
+      var allUsers = response.users;
+      // Empty the anchor
+      var allUsersAnchor = document.getElementById("all-users-anchor");
+      allUsersAnchor.innerHTML = "";
+      // Append users to anchor
+      allUsers.forEach((user) => {
+        allUsersAnchor.innerHTML += getUserDisplayEle(user);
+      });
+    });
 }
 
 function getUserDisplayEle(user) {
+  if (user.role == 1) {
     return ` <tr>
     <td>${user.id}</td>
     <td>${user.name}</td>
     <td>${user.email}</td>
-    <td>${user.role}</td>
+    <td>admin</td>
     <td> <a class="fa fa-trash p-4 delete-user-btn" onclick="deleteUser(${user.id})">
 </a></td>
 </tr>`;
+  } else {
+    return ` <tr>
+<td>${user.id}</td>
+<td>${user.name}</td>
+<td>${user.email}</td>
+<td>Employee</td>
+<td> <a class="fa fa-trash p-4 delete-user-btn" onclick="deleteUser(${user.id})">
+</a></td>
+</tr>`;
+  }
 }
 
 /******************************************************************************
@@ -41,28 +52,28 @@ const addBtn = document.getElementById("add-btn");
 addBtn.addEventListener("click", () => addUserOne(), false);
 
 function addUserOne() {
-    var nameInput = document.getElementById("name-input");
-    var emailInput = document.getElementById("email-input");
-    var pwdlInput = document.getElementById("pwd-input");
-    var roleInput = document.getElementById("role-input");
-    var data = {
-        user: {
-            name: nameInput.value,
-            email: emailInput.value,
-            pwdHash: pwdlInput.value,
-            role: parseInt(roleInput.value),
-        },
-    };
-    Http.Post("/api/users/add", data).then(() => {
-        window.location.href = "/users";
-    });
+  var nameInput = document.getElementById("name-input");
+  var emailInput = document.getElementById("email-input");
+  var pwdlInput = document.getElementById("pwd-input");
+  var roleInput = document.getElementById("role-input");
+  var data = {
+    user: {
+      name: nameInput.value,
+      email: emailInput.value,
+      pwdHash: pwdlInput.value,
+      role: parseInt(roleInput.value),
+    },
+  };
+  Http.Post("/api/users/add", data).then(() => {
+    window.location.href = "/users";
+  });
 }
 
 function deleteUser(id) {
-    var id = id;
-    Http.Delete("/api/users/delete/" + id).then(() => {
-        displayUsers();
-    });
+  var id = id;
+  Http.Delete("/api/users/delete/" + id).then(() => {
+    displayUsers();
+  });
 }
 
 /******************************************************************************
@@ -70,15 +81,15 @@ function deleteUser(id) {
  ******************************************************************************/
 
 function logoutUser() {
-    Http.Get("/api/auth/logout").then(() => {
-        window.location.href = "/";
-    });
+  Http.Get("/api/auth/logout").then(() => {
+    window.location.href = "/";
+  });
 }
 
 function hidenav() {
-    let role = localStorage.getItem("role");
+  let role = localStorage.getItem("role");
 
-    if (role == 2) {
-        window.location.href = "/Books";
-    }
+  if (role == 2) {
+    window.location.href = "/Books";
+  }
 }
